requirements.txt
Package               Version
--------------------- -------
faiss-cpu             1.13.0
langchain             0.1.20
langchain-groq        0.1.5
pip                   25.3
pypdf                 6.4.0
python-dotenv         1.2.1
sentence-transformers 3.1.1
setuptools            80.9.0
